+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 
+# v4.x
+- Build artifacts for NVList projects are now stored relative to the project folder (`vnRoot/build-out`), instead of inside the engine folder (`build-tools`).
+
 # v4.3.0
 - Added a global preference (`fullscreen`) to choose whether to start in full-screen mode or windows mode. Default value is `true`, unless `debug` mode is enabled.
 - Improvements to text rendering. In particular, glyphs are now rendered slightly thicker which really helps with thin lines in some fonts.